add delete comment by org name

+++ pkg/core/service/comment/getlist.go
@@ -12,8 +12,8 @@ type GetListReq struct {
 	ID             uint64
 	OrganizationId uint64
 
-	Limit  int
-	Offset int
+	Limit  int64
+	Offset int64
 }
 
 type GetListResp struct {
